+++ app/models/user.rb
@@ -39,7 +39,8 @@ class User < ApplicationRecord
   multisearchable against: [:full_name, :email, :username, :keywords, :phone, :role]
 
   # Scopes
-  scope :editors, -> { where(editor: true).or(where(admin: true)) }
+  scope :admins, -> { current.where(admin: true) }
+  scope :editors, -> { current.where(editor: true).or(admins) }
 
   # Validations
   validates :full_name, presence: true
@@ -72,6 +73,7 @@ class User < ApplicationRecord
 
   def check_approval_email
     return unless approved? && approval_sent_at.nil?
+
     update(approval_sent_at: Time.zone.now)
     send_approval_email!
   end
@@ -88,22 +90,26 @@ class User < ApplicationRecord
   # Override Devise built-in password reset notification email method
   def send_reset_password_instructions
     return unless EMAILS_ENABLED && !deleted?
+
     super
   end
 
   # Override Devise built-in unlock instructions notification email method
   def send_unlock_instructions
     return unless EMAILS_ENABLED && !deleted?
+
     super
   end
 
   def send_confirmation_instructions
     return unless EMAILS_ENABLED && !deleted? && !disposable_email?
+
     super
   end
 
   def send_on_create_confirmation_instructions
     return unless EMAILS_ENABLED && !deleted? && !disposable_email?
+
     send_welcome_email!
   end
 
@@ -113,8 +119,18 @@ class User < ApplicationRecord
 
   def new_registration!
     return unless EMAILS_ENABLED
-    User.current.where(admin: true).find_each do |admin|
-      RegistrationMailer.user_registered(admin, self).deliver_later
+
+    User.admins.find_each do |admin|
+      RegistrationMailer.account_registered(admin, self).deliver_later
+    end
+  end
+
+  # Override Devise::Confirmable#after_confirmation
+  def after_confirmation
+    return unless EMAILS_ENABLED
+
+    User.admins.find_each do |admin|
+      RegistrationMailer.account_confirmed(admin, self).deliver_later
     end
   end